bakup

+++ create_script/translate_file.py
@@ -1,28 +1,29 @@
-import translate
 import os
-import translate
+import create_script.translating
 
-file = ""
-finish = False
 def translate_lines(path = ""):
-    global finish
-    filelocal = ""
-    with open(path, 'r') as archive:
-        lines = archive.readlines()
-        
-        for line in lines:
-            if(line.find("@")<=0):
-                text=line.replace("|\n","").replace("|","").replace("\n","")
-                translated = translate.english_to_portuguese(text)
-                #finish = False
+    conteudo = os.listdir(path)
+    for item in conteudo:
+        print(item)
+        item_caminho = os.path.join(path, item)
+        if os.path.isfile(item_caminho):
+            filelocal = ""
+            with open(item_caminho, 'r', encoding='ASCII') as archive:
+                lines = archive.readlines()
+    
+                for line in lines:
+                    if(line.find("@")<=0):
+                        text=line.replace("|\n","").replace("|","").replace("\n","")
+                        translated = create_script.translating.english_to_portuguese(text)
+                        #finish = False
 
-            if(len(translated)>0):
-                line = f"{text}|{translated}@\n"
+                    if(len(translated)>0):
+                        line = f"{text}|{translated}@\n"
 
-            filelocal += line
-            print(line)
+                    filelocal += line
+                    print(line)
 
-    with open(path, 'w') as newFile:
-        newFile.write(filelocal)
+            with open(path, 'w') as newFile:
+                newFile.write(filelocal)
 
     return True
